+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.1.8] - Unreleased
+### Added
+- Types for SVG elements
+
 ### Fixed
 - Performance: removed unnecessary regular expressions.
+- Updated HTML & CSS types
 
 ## [0.1.7] - 2025-03-21
 ### Fixed